Wrapping the profile component in a private route, so if there is no user logged in, they get redirected back to home. Pretty cool, makes the app functional in an anticipated error situation, prevents the user from being on a page they shouldn't be on.
For the private route we created a custom hook (to use elsewhere in the app if we need) which returns whether a user is logged in and if the app is loading.

We also clean up a potential memory leak in the hook from the useEffect.

Adding a spinner, pretty cool, all it does it spin, then we control when to show the spinner component.