installed uuid. Got it to save photos to firebase storage, sweet.
StoreImage() is used to iterate through all images. The function creates a promise, uploads the images, logs the progress, and resolves the promise when the downloadURL for the uploaded images are retrieved.

We use promise.all as we're mapping through each image. Remember storeImage is a function that creates a promise.